Fixed an off by one error in the clock divider.
I was starting from 0, but then resetting on N leading to a
cycle of length N+1.

I fixed this by subtracting one from the pulseEveryNCycles input. This
might be a little inefficient of hardware.  I don't know if I'm better
off creating a register for this instead of using subtraction logic.